<feat(Benefits)>: "일급 컬렉션 Benefits 생성" - 각 혜택을 List로 다루는 일급 컬렉션 생성 - 정적 팩토리를 통해 미적용 시와 적용 시로 생성하도록 구현 - 기능 목록 수정

+++ src/main/java/christmas/domain/menu/Menus.java
@@ -33,14 +33,14 @@ public class Menus {
                 .toList();
     }
 
-    public long calculateMainMenuNum() {
-        return menus.keySet().stream()
+    public int calculateMainMenuNum() {
+        return (int) menus.keySet().stream()
                 .filter(key -> key.getMenuCode() == 2)
                 .count();
     }
 
-    public long calculateDessertMenuNum() {
-        return menus.keySet().stream()
+    public int calculateDessertMenuNum() {
+        return (int) menus.keySet().stream()
                 .filter(key -> key.getMenuCode() == 3)
                 .count();
     }
